Make generation of SCM snapshots and tarballs optional Patch on 5.1 originally from Roland Mas

+++ src/plugins/scmarch/common/ArchPlugin.class.php
@@ -74,6 +74,12 @@ class ArchPlugin extends SCMPlugin {
 	}
 
 	function generateSnapshots ($params) {
+		$us = forge_get_config('use_scm_snapshots') ;
+		$ut = forge_get_config('use_scm_tarballs') ;
+		if (!$us && !$ut) {
+			return false ;
+		}
+
 		$project = $this->checkParams ($params) ;
 		if (!$project) {
 			return false ;
@@ -107,11 +113,13 @@ class ArchPlugin extends SCMPlugin {
 		if ($tmp == '') {
 			return false ;
 		}
-		system ("tar cCf $toprepo - ".$project->getUnixName() ."|".forge_get_config('compression_method')."> $tmp/tarball") ;
-		chmod ("$tmp/tarball", 0644) ;
-		copy ("$tmp/tarball", $tarball) ;
-		unlink ("$tmp/tarball") ;
-		system ("rm -rf $tmp") ;
+		if ($ut) {
+			system ("tar cCf $toprepo - ".$project->getUnixName() ."|".forge_get_config('compression_method')."> $tmp/tarball") ;
+			chmod ("$tmp/tarball", 0644) ;
+			copy ("$tmp/tarball", $tarball) ;
+			unlink ("$tmp/tarball") ;
+			system ("rm -rf $tmp") ;
+		}
 	}
   }
