+++ src/plugins/scmbzr/common/BzrPlugin.class.php
@@ -453,6 +453,11 @@ class BzrPlugin extends SCMPlugin {
 	}
 
 	function generateSnapshots ($params) {
+		$us = forge_get_config('use_scm_snapshots') ;
+		$ut = forge_get_config('use_scm_tarballs') ;
+		if (!$us && !$ut) {
+			return false ;
+		}
 
 		$project = $this->checkParams ($params) ;
 		if (!$project) {
@@ -487,23 +492,27 @@ class BzrPlugin extends SCMPlugin {
 		if ($tmp == '') {
 			return false ;
 		}
-		$today = date ('Y-m-d') ;
-		$branch = $this->findMainBranch ($project) ;
-		if ($branch != '') {
-			system ("bzr export --root=$group_name-scm-$today --format=tar - $repo/$bname |".forge_get_config('compression_method')."> $tmp/snapshot") ;
-			chmod ("$tmp/snapshot", 0644) ;
-			copy ("$tmp/snapshot", $snapshot) ;
-			unlink ("$tmp/snapshot") ;
-			system ("rm -rf $tmp/$dir") ;
-		} else {
-			unlink ($snapshot) ;
+		if ($us) {
+			$today = date ('Y-m-d') ;
+			$branch = $this->findMainBranch ($project) ;
+			if ($branch != '') {
+				system ("bzr export --root=$group_name-scm-$today --format=tar - $repo/$bname |".forge_get_config('compression_method')."> $tmp/snapshot") ;
+				chmod ("$tmp/snapshot", 0644) ;
+				copy ("$tmp/snapshot", $snapshot) ;
+				unlink ("$tmp/snapshot") ;
+				system ("rm -rf $tmp/$dir") ;
+			} else {
+				unlink ($snapshot) ;
+			}
 		}
 
-		system ("tar cCf $toprepo - ".$project->getUnixName() ."|".forge_get_config('compression_method')."> $tmp/tarball") ;
-		chmod ("$tmp/tarball", 0644) ;
-		copy ("$tmp/tarball", $tarball) ;
-		unlink ("$tmp/tarball") ;
-		system ("rm -rf $tmp") ;
+		if ($ut) {
+			system ("tar cCf $toprepo - ".$project->getUnixName() ."|".forge_get_config('compression_method')."> $tmp/tarball") ;
+			chmod ("$tmp/tarball", 0644) ;
+			copy ("$tmp/tarball", $tarball) ;
+			unlink ("$tmp/tarball") ;
+			system ("rm -rf $tmp") ;
+		}
 	}
   }
