+++ src/plugins/scmdarcs/common/DarcsPlugin.class.php
@@ -381,6 +381,11 @@ class DarcsPlugin extends SCMPlugin {
 	}
 
 	function generateSnapshots ($params) {
+		$us = forge_get_config('use_scm_snapshots') ;
+		$ut = forge_get_config('use_scm_tarballs') ;
+		if (!$us && !$ut) {
+			return false ;
+		}
 
 		$project = $this->checkParams ($params) ;
 		if (!$project) {
@@ -414,21 +419,25 @@ class DarcsPlugin extends SCMPlugin {
 		if ($tmp == '') {
 			return false ;
 		}
-		$today = date ('Y-m-d') ;
-		$dir = $project->getUnixName ()."-$today" ;
-		system ("mkdir -p $tmp/$dir") ;
-		system ("cd $tmp ; darcs $repo $dir > /dev/null 2>&1") ;
-		system ("tar cCf $tmp - $dir |".forge_get_config('compression_method')."> $tmp/snapshot") ;
-		chmod ("$tmp/snapshot", 0644) ;
-		copy ("$tmp/snapshot", $snapshot) ;
-		unlink ("$tmp/snapshot") ;
-		system ("rm -rf $tmp/$dir") ;
+		if ($us) {
+			$today = date ('Y-m-d') ;
+			$dir = $project->getUnixName ()."-$today" ;
+			system ("mkdir -p $tmp/$dir") ;
+			system ("cd $tmp ; darcs $repo $dir > /dev/null 2>&1") ;
+			system ("tar cCf $tmp - $dir |".forge_get_config('compression_method')."> $tmp/snapshot") ;
+			chmod ("$tmp/snapshot", 0644) ;
+			copy ("$tmp/snapshot", $snapshot) ;
+			unlink ("$tmp/snapshot") ;
+			system ("rm -rf $tmp/$dir") ;
+		}
 
-		system ("tar cCf $toprepo - ".$project->getUnixName() ."|".forge_get_config('compression_method')."> $tmp/tarball") ;
-		chmod ("$tmp/tarball", 0644) ;
-		copy ("$tmp/tarball", $tarball) ;
-		unlink ("$tmp/tarball") ;
-		system ("rm -rf $tmp") ;
+		if ($ut) {
+			system ("tar cCf $toprepo - ".$project->getUnixName() ."|".forge_get_config('compression_method')."> $tmp/tarball") ;
+			chmod ("$tmp/tarball", 0644) ;
+			copy ("$tmp/tarball", $tarball) ;
+			unlink ("$tmp/tarball") ;
+			system ("rm -rf $tmp") ;
+		}
 	}
 
 	function gatherStats ($params) {
